기타 css 수정 및 코드 추가

+++ ShoppingMall/src/main/java/com/lkm/shoppingmall/dto/searchDto.java
@@ -28,8 +28,8 @@ public class searchDto {
 		this.pName = pName;
 	}
 
-	public double getrRating() {
-		return rRating;
+	public int getrRating() {
+		return (int) rRating;
 	}
 
 	public void setrRating(double rRating) {
